word list: simplify merge script
Concatenating the files and then sorting them can be done by `sort(1)`
itself, and is more robust than using a loop.

The script could also write directily to `words.txt`, but if the `sort`
command fails for whatever reason, we would end up with an empty file
(the `>` truncates the file before `sort` even starts). So we continue
to use a temporary file and move it over the existsing `words.txt` only
if the `sort` command works correctly.

Also use `mktmp(1)` to create the file, which interprets/respects `$TMP`
and creates a random file name, to avoid symlink attacks. It's not a
high-risk scenario, but still good practice. We use `trap ... EXIT` to
make sure the temporary file is cleaned up even if `sort` or `mv` fail.